redux refactor complete
closes #733
