+++ src/components/WalletSelect/tabs/Txns.js
@@ -9,7 +9,7 @@ import { usePool } from '../../../store/pool'
 import { useSparta } from '../../../store/sparta'
 import { useSynth } from '../../../store/synth'
 import { useRouter } from '../../../store/router'
-import { useWeb3 } from '../../../store/web3/selector'
+import { useWeb3 } from '../../../store/web3'
 import { formatFromWei } from '../../../utils/bigNumber'
 import { getPool, getSynth, getToken } from '../../../utils/math/utils'
 import spartaLpIcon from '../../../assets/tokens/sparta-lp.svg'
