--- src/store/helpers.js
@@ -1,9 +1,0 @@
-export const errorToDispatch = (type, error) => ({
-  type,
-  error,
-})
-
-export const payloadToDispatch = (type, payload) => ({
-  type,
-  payload,
-})
