--- src/store/utils/selector.js
@@ -1,3 +1,0 @@
-import { useSelector } from 'react-redux'
-
-export const useUtils = () => useSelector((state) => state.utils)
