--- src/store/web3/selector.js
@@ -1,3 +1,0 @@
-import { useSelector } from 'react-redux'
-
-export const useWeb3 = () => useSelector((state) => state.web3)
