--- src/store/web3/types.js
@@ -1,12 +1,0 @@
-export const ADD_NETWORK_MM = 'ADD_NETWORK_MM'
-export const ADD_NETWORK_BC = 'ADD_NETWORK_BC'
-export const WEB3_TXN = 'WEB3_TXN'
-export const GET_ALLOWANCE1 = 'GET_ALLOWANCE1'
-export const GET_ALLOWANCE2 = 'GET_ALLOWANCE2'
-export const WATCH_ASSET = 'WATCH_ASSET'
-export const SPARTA_PRICE = 'SPARTA_PRICE'
-export const WEB3_LOADING = 'WEB3_LOADING'
-export const WEB3_ERROR = 'WEB3_ERROR'
-export const EVENT_ARRAY = 'EVENT_ARRAY'
-export const RPC_BLOCKS = 'RPC_BLOCKS'
-export const WEB3_METRICS = 'WEB3_METRICS'
